specialty_coach/requirements.txt
langchain
langchain-google-genai
python-dotenv
PyPDF2
